Add files via upload
Abstract— Direct marketing has generated an increasing interest among academics and practitioners over the past few years. This is due to competitive market environment, advancement in technology and changing behavior of customers which are difficult to predict. The   size   of   these data   sources is impossible  for  a  human  analyst  to  come  up  with  interesting information  that  will  help  in  the  decision-making  process. Data    mining    models    are    completely    helping    in    the performance   of   these   campaigns. The primary focus of the paper is to study the data and build a predictive model to predict whether a customer will go for term deposit or not by using the attributes. The classification goal is to predict if the client will subscribe a term deposit (variable y). Such model can increase campaign efficiency by identifying the main characteristics that affect success, helping in a better management of the available resources (e.g. human effort, phone calls, time) and selection of a high quality and affordable set of potential buying customers.